Update s04.rpy
In German we often use the English words. Like Bitch, and Pussy. And words like Pussy can be translated in 3-4 various Words. So I would recommend that you just Stick to English Words for these kind of words:
Examples: Cumshot - Abwichsen, Wixen, Abbspritzen, Gesichtsdusche All can be used for this Word, so best Way Stick with Fetish words to the english. All will know the meaning and it is better then uneven Translations.